refactor: Rename user dashboard file
