+++ frontend/js/tests/common/listens/ListensControls.test.tsx
@@ -7,12 +7,12 @@ import {
   within,
 } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { renderWithProviders } from "../test-utils/rtl-test-utils";
-import * as recentListensPropsOneListen from "../__mocks__/recentListensPropsOneListen.json";
-import * as recentListensPropsThreeListens from "../__mocks__/recentListensPropsThreeListens.json";
-import * as getFeedbackByMsidResponse from "../__mocks__/getFeedbackByMsidResponse.json";
-import * as getMultipleFeedbackResponse from "../__mocks__/getMultipleFeedbackResponse.json";
-import Listens, { ListensProps } from "../../src/user/Dashboard";
+import { renderWithProviders } from "../../test-utils/rtl-test-utils";
+import * as recentListensPropsOneListen from "../../__mocks__/recentListensPropsOneListen.json";
+import * as recentListensPropsThreeListens from "../../__mocks__/recentListensPropsThreeListens.json";
+import * as getFeedbackByMsidResponse from "../../__mocks__/getFeedbackByMsidResponse.json";
+import * as getMultipleFeedbackResponse from "../../__mocks__/getMultipleFeedbackResponse.json";
+import Listens, { ListensProps } from "../../../src/user/Dashboard";
 
 // Font Awesome generates a random hash ID for each icon everytime.
 // Mocking Math.random() fixes this
